fifth commit

+++ database/migrations/2025_09_17_080952_create_orders_table.php
@@ -10,7 +10,7 @@ return new class extends Migration {
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->decimal('total_price', 12, 2);
-            $table->tinyInteger('status')->default(0);
+            $table->smallInteger('status')->default(0);
             // status: 0 = pending, 1 = paid, 2 = shipped, 3 = completed, 4 = cancelled
             $table->timestamps();
         });
